requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
diff-match-patch==20241021
Django==5.2.5
django-crispy-forms==2.4
django-environ==0.12.0
django-guardian==3.0.3
django-import-export==4.3.9
django-simple-history==3.10.1
django-tailwind==4.2.0
django-unfold==0.64.2
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250809
tzdata==2025.2
urllib3==2.5.0

# AI and Machine Learning
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
python-dotenv==1.0.0
